+++ src/components/show-more.js
@@ -1,0 +1,7 @@
+/**
+ * создаёт и возвращает разметку кнопки 'Показать больше'
+ * @return {String}
+ */
+export const getButtonShowMoreTemplate = () => (/* html */
+  `<button class="films-list__show-more">Show more</button>`
+);
