+++ src/data/film-details.js
@@ -1,0 +1,70 @@
+const filmDetailsDescriptionData = [
+  {
+    term: `Director`,
+    cell: `Anthony Mann`
+  },
+  {
+    term: `Writers`,
+    cell: `Anne Wigton, Heinz Herald, Richard Weil`
+  },
+  {
+    term: `Actors`,
+    cell: `Erich von Stroheim, Mary Beth Hughes, Dan Duryea`
+  },
+  {
+    term: `Release Date`,
+    cell: `30 March 1945`
+  },
+  {
+    term: `Runtime`,
+    cell: `1h 18m`
+  },
+  {
+    term: `Country`,
+    cell: `USA`
+  },
+  {
+    term: `Genres`,
+    cell: `Drama`
+  }
+];
+
+const filmDetailsControlsData = [
+  {
+    name: `watchlist`,
+    desc: `Add to watchlist`
+  },
+  {
+    name: `watched`,
+    desc: `Already watched`
+  },
+  {
+    name: `favorite`,
+    desc: `Add to favorites`
+  }
+];
+
+const FilmDetailsEmojiData = [
+  {
+    id: `emoji-smile`,
+    value: `sleeping`,
+    imgSrc: `./images/emoji/smile.png`
+  },
+  {
+    id: `emoji-sleeping`,
+    value: `neutral-face`,
+    imgSrc: `./images/emoji/sleeping.png`
+  },
+  {
+    id: `emoji-gpuke`,
+    value: `grinning`,
+    imgSrc: `./images/emoji/puke.png`
+  },
+  {
+    id: `emoji-angry`,
+    value: `grinning`,
+    imgSrc: `./images/emoji/angry.png`
+  }
+];
+
+export {filmDetailsDescriptionData, filmDetailsControlsData, FilmDetailsEmojiData};
