+++ src/data/menu-links.js
@@ -1,0 +1,40 @@
+export const menuLinksData = [
+  {
+    name: `All movies`,
+    path: `#all`,
+    active: true,
+    additional: false,
+    desription: false
+  },
+  {
+    name: `Watchlist`,
+    path: `#watchlist`,
+    active: false,
+    additional: false,
+    desription: true,
+    desriptionItemCount: 13
+  },
+  {
+    name: `History`,
+    path: `#history`,
+    active: false,
+    additional: false,
+    desription: true,
+    desriptionItemCount: 8
+  },
+  {
+    name: `Favorites`,
+    path: `#favorites`,
+    active: false,
+    additional: false,
+    desription: true,
+    desriptionItemCount: 4
+  },
+  {
+    name: `Stats`,
+    path: `#stats`,
+    active: false,
+    additional: true,
+    desription: false
+  }
+];
