+++ src/data/sort-links.js
@@ -1,0 +1,14 @@
+export const sortLinksData = [
+  {
+    name: `Sort by default`,
+    active: true
+  },
+  {
+    name: `Sort by date`,
+    active: false
+  },
+  {
+    name: `Sort by rating`,
+    active: false
+  }
+];
